Move checking of CV history changed since time to `PatientRepository`
