+++ app/src/main/java/org/simple/clinic/patient/PatientRepository.kt
@@ -636,5 +636,18 @@ class PatientRepository @Inject constructor(
         )
   }
 
+  @Suppress("DeprecatedCallableAddReplaceWith")
+  @Deprecated(message = "temporarily added for refactoring, remove later")
+  // TODO(vs): 2019-11-27 Remove this once the refactoring to move data changed since time method to the patient repository
+  fun hasMedicalHistoryForPatientChangedSince(patientUuid: UUID, instant: Instant): Boolean {
+    return database
+        .medicalHistoryDao()
+        .hasMedicalHistoryForPatientChangedSince(
+            patientUuid = patientUuid,
+            instantToCompare = instant,
+            pendingStatus = PENDING
+        )
+  }
+
   private data class BusinessIdMetaAndVersion(val metaData: String, val metaDataVersion: MetaDataVersion)
 }
